SdmForm():
$formElements property now defaults to an empty array.

Created new method sdmFormUseDefaultFormElements() which assigns the default form elements to the $formElements property when it's called.

Added new method sdmFormDetermineRootUrl() which is used by sdmFormBuildForm() and sdmFormOpenForm() to determine root url if not provided via the $rootUrl parameter for each method.

Fixed bug in sdmFormBuildFormElements() that was causing individual form element html to be appended to other form element's html as the elements were constructed.

Added new methods sdmFormOpenForm() and sdmFormCloseForm() which can be used to create the opening and closing form html when building form manually with sdmFormGetFormElementHtml().

Added new method sdmFormCreateFormElement() which can be used to create individual form elements. sdmFormCreateFormElement() will add the form elements it creates to the SdmForm()'s $formElements property automatically.

Created new method sdmFormRecursiveInArray() which provides a way to recursively check for the existence of a value in a multi-dimensional array.